creating a second auth verification
